+++ Changelog.md
@@ -1,5 +1,9 @@
 AptioFix Changelog
 ==================
 
-#### v1.0
-- Initial release
+#### R22
+- Fixed cdboot.efi conflict
+- Added a workaround to Windows Virtualization-based Security
+
+#### R21
+- First binary release